MNT: move Artistlist into artist module
