+++ doc/api/axes_api.rst
@@ -634,5 +634,3 @@ Other
    Axes.get_figure
    Axes.figure
    Axes.remove
-
-.. autoclass:: matplotlib.axes.Axes.ArtistList
