+++ doc/api/next_api_changes/deprecations/31794-REC.rst
@@ -1,0 +1,3 @@
+The ``Axes.ArtistList`` attribute
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+... is deprecated.  Use `.artist.ArtistList` instead.
