+++ lib/matplotlib/artist.pyi
@@ -14,7 +14,7 @@ from .transforms import (
 
 import numpy as np
 
-from collections.abc import Callable, Iterable
+from collections.abc import Callable, Iterable, Iterator, Sequence
 from typing import Any, Literal, NamedTuple, TextIO, overload
 from numpy.typing import ArrayLike
 
@@ -186,6 +186,36 @@ class ArtistInspector:
     def pprint_setters_rest(self, prop: str, leadingspace: int = ...) -> str: ...
     def properties(self) -> dict[str, Any]: ...
     def pprint_getters(self) -> list[str]: ...
+
+
+class ArtistList[T: Artist](Sequence[T]):
+    def __init__(
+        self,
+        axes: _AxesBase,
+        prop_name: str,
+        valid_types: type | Iterable[type] | None = ...,
+        invalid_types: type | Iterable[type] | None = ...,
+    ) -> None: ...
+    def __len__(self) -> int: ...
+    def __iter__(self) -> Iterator[T]: ...
+    @overload
+    def __getitem__(self, key: int) -> T: ...
+    @overload
+    def __getitem__(self, key: slice) -> list[T]: ...
+
+    @overload
+    def __add__(self, other: ArtistList[T]) -> list[T]: ...
+    @overload
+    def __add__(self, other: list[Any]) -> list[Any]: ...
+    @overload
+    def __add__(self, other: tuple[Any]) -> tuple[Any]: ...
+
+    @overload
+    def __radd__(self, other: ArtistList[T]) -> list[T]: ...
+    @overload
+    def __radd__(self, other: list[Any]) -> list[Any]: ...
+    @overload
+    def __radd__(self, other: tuple[Any]) -> tuple[Any]: ...
 
 def getp(obj: Artist, property: str | None = ...) -> Any: ...
