+++ lib/matplotlib/axes/_base.pyi
@@ -1,9 +1,9 @@
 import matplotlib.artist as martist
 
 import datetime
-from collections.abc import Callable, Iterable, Iterator, Sequence
+from collections.abc import Callable, Iterable, Sequence
 from matplotlib import cbook
-from matplotlib.artist import Artist
+from matplotlib.artist import Artist, ArtistList
 from matplotlib.axes import Axes
 from matplotlib.axis import Axis, XAxis, YAxis, Tick
 from matplotlib.backend_bases import RendererBase, MouseButton, MouseEvent
@@ -134,49 +134,20 @@ class _AxesBase(martist.Artist):
     def clear(self) -> None: ...
     def cla(self) -> None: ...
 
-    class ArtistList[T: Artist](Sequence[T]):
-        def __init__(
-            self,
-            axes: _AxesBase,
-            prop_name: str,
-            valid_types: type | Iterable[type] | None = ...,
-            invalid_types: type | Iterable[type] | None = ...,
-        ) -> None: ...
-        def __len__(self) -> int: ...
-        def __iter__(self) -> Iterator[T]: ...
-        @overload
-        def __getitem__(self, key: int) -> T: ...
-        @overload
-        def __getitem__(self, key: slice) -> list[T]: ...
-
-        @overload
-        def __add__(self, other: _AxesBase.ArtistList[T]) -> list[T]: ...
-        @overload
-        def __add__(self, other: list[Any]) -> list[Any]: ...
-        @overload
-        def __add__(self, other: tuple[Any]) -> tuple[Any]: ...
-
-        @overload
-        def __radd__(self, other: _AxesBase.ArtistList[T]) -> list[T]: ...
-        @overload
-        def __radd__(self, other: list[Any]) -> list[Any]: ...
-        @overload
-        def __radd__(self, other: tuple[Any]) -> tuple[Any]: ...
-
     @property
-    def artists(self) -> _AxesBase.ArtistList[Artist]: ...
+    def artists(self) -> ArtistList[Artist]: ...
     @property
-    def collections(self) -> _AxesBase.ArtistList[Collection]: ...
+    def collections(self) -> ArtistList[Collection]: ...
     @property
-    def images(self) -> _AxesBase.ArtistList[AxesImage]: ...
+    def images(self) -> ArtistList[AxesImage]: ...
     @property
-    def lines(self) -> _AxesBase.ArtistList[Line2D]: ...
+    def lines(self) -> ArtistList[Line2D]: ...
     @property
-    def patches(self) -> _AxesBase.ArtistList[Patch]: ...
+    def patches(self) -> ArtistList[Patch]: ...
     @property
-    def tables(self) -> _AxesBase.ArtistList[Table]: ...
+    def tables(self) -> ArtistList[Table]: ...
     @property
-    def texts(self) -> _AxesBase.ArtistList[Text]: ...
+    def texts(self) -> ArtistList[Text]: ...
     def get_facecolor(self) -> ColorType: ...
     def set_facecolor(self, color: ColorType | None) -> None: ...
     @overload
